sockatz: refactor methods into package client
